backend/requirements.txt
fastapi
uvicorn
requests
google-auth
google-cloud-logging
python-multipart
google-cloud-aiplatform>=1.38.0

